+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Added
+
+- `network set_max_communities <network> <max_communities>` command to set the maximum number of communities allowed on a network.
+- `network unset_max_communities <network>` command to unset the maximum number of communities allowed on a network.
+
+### Changed
+
+- Renamed `network policy_set_prefix` to `network policy_set_pattern`.
+- Renamed `network policy_unset_prefix` to `network policy_unset_pattern`.
 
 ## [1.7.3](https://github.com/unioslo/mreg-cli/releases/tag/1.7.3) - 2025-11-18
 
@@ -83,6 +93,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `logging start` no longer toggles logging on and off. Now aborts if logging is already enabled.
 - `logging stop` now aborts if logging is not enabled.
 - `logging status` now shows if logging is enabled or not.
+- `network policy_set_prefix` renamed to `network policy_set_pattern`.
+- `network policy_unset_prefix` renamed to `network policy_unset_pattern`.
 
 ## [1.6.0](https://github.com/unioslo/mreg-cli/releases/tag/1.6.0) - 2025-10-13